+++ src/main.ts
@@ -6,6 +6,7 @@ import log from 'loglevel'
 
 import VueToastificationPlugin from 'vue-toastification'
 import { ICON_NAMES, ROUTE_NAMES } from '@/enums'
+import VueTheMask from 'vue-the-mask'
 import { createApp, getCurrentInstance, h } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { i18n } from '@/localization'
@@ -27,11 +28,19 @@ const app = createApp({
 
 log.setDefaultLevel(config.LOG_LEVEL)
 
-app.use(router).use(store).use(i18n).use(VueToastificationPlugin)
+app
+  .use(router)
+  .use(store)
+  .use(i18n)
+  .use(VueToastificationPlugin)
+  .use(VueTheMask)
 
 app.config.globalProperties.$routes = ROUTE_NAMES
 app.config.globalProperties.$config = config
 app.config.globalProperties.$icons = ICON_NAMES
+app.config.globalProperties.$filter = {
+  username: (value: string) => '@' + value,
+}
 
 app.config.errorHandler = function (err, vm, info) {
   log.error(`Error: ${err}; Info: ${info}`)