+++ src/pages/UnverifiedUsers.vue
@@ -31,17 +31,19 @@
       ref="usersList"
     />
 
-    <create-user-modal
-      :is-shown="isShownCreateUserModal"
-      @submit="reloadCreateUserModal"
-      @cancel="toggleCreateUserModal"
-    />
+    <transition-modal>
+      <create-user-modal
+        v-if="isShownCreateUserModal"
+        @submit="reloadCreateUserModal"
+        @cancel="toggleCreateUserModal"
+      />
+    </transition-modal>
   </div>
 </template>
 
 <script lang="ts" setup>
 import { computed, ref } from 'vue'
-import { AppButton, CreateUserModal } from '@/common'
+import { AppButton, CreateUserModal, TransitionModal } from '@/common'
 import { InputField } from '@/fields'
 import UnverifiedUsersList from '@/pages/unverified-users/UnverifiedUsersList.vue'
 import { ICON_NAMES } from '@/enums'
@@ -87,7 +89,7 @@ const clearInput = () => {
 
 .unverified-users__actions {
   display: flex;
-  gap: toRem(10);
+  gap: toRem(23);
 }
 
 .unverified-users__search-input {